started moving batch operations on arrays of mxCells to mxCells.ts from mxGraphModel.ts
